requirements.txt
# --- Common ---
pydub

# --- Online mode (Groq API) ---
groq
python-dotenv

# --- Local mode (Whisper) ---
openai-whisper

# The following line is an EXAMPLE for GPU installation (local mode).
# It's better to run this command separately after checking your CUDA version.
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121